Removed dup logging code
