+++ src/test/java/elegit/LocalSshAuthenticationTests.java
@@ -7,6 +7,7 @@ import elegit.models.AuthMethod;
 import elegit.models.ClonedRepoHelper;
 import elegit.models.ExistingRepoHelper;
 import elegit.models.RepoHelper;
+import elegit.sshauthentication.DetailedSshLogger;
 import elegit.sshauthentication.ElegitUserInfoTest;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -85,30 +86,10 @@ public class LocalSshAuthenticationTests {
         }
     }
 
-    // http://www.jcraft.com/jsch/examples/Logger.java.html
-    public static class MyLogger implements com.jcraft.jsch.Logger {
-        static java.util.Hashtable<Integer,String> name=new java.util.Hashtable<>();
-        static{
-            name.put(DEBUG, "DEBUG: ");
-            name.put(INFO, "INFO: ");
-            name.put(WARN, "WARN: ");
-            name.put(ERROR, "ERROR: ");
-            name.put(FATAL, "FATAL: ");
-        }
-        public boolean isEnabled(int level){
-            return true;
-        }
-        public void log(int level, String message){
-            System.err.print(name.get(level));
-            System.err.println(message);
-        }
-    }
-
-
     @Test
     public void testSshPrivateKey() throws Exception {
         // Uncomment this to get detail SSH logging info, for debugging
-        //JSch.setLogger(new MyLogger());
+//        JSch.setLogger(new DetailedSshLogger());
 
         // Set up test SSH server.
         sshd = SshServer.setUpDefaultServer();
