+++ src/test/java/elegitfx/SshPrivateKeyPasswordCloneFXTest.java
@@ -13,6 +13,7 @@ import elegit.models.ExistingRepoHelper;
 import elegit.models.RepoHelper;
 import elegit.models.SessionModel;
 import elegit.monitors.RepositoryMonitor;
+import elegit.sshauthentication.DetailedSshLogger;
 import elegit.sshauthentication.ElegitUserInfoTest;
 import elegit.treefx.Cell;
 import javafx.fxml.FXMLLoader;
@@ -55,6 +56,7 @@ import org.junit.Test;
 import org.junit.rules.TestName;
 import org.loadui.testfx.GuiTest;
 import org.testfx.framework.junit.ApplicationTest;
+import sharedrules.TestUtilities;
 import sharedrules.TestingLogPathRule;
 import sharedrules.TestingRemoteAndLocalReposRule;
 
@@ -163,27 +165,6 @@ public class SshPrivateKeyPasswordCloneFXTest extends ApplicationTest {
 
     }
 
-
-    // http://www.jcraft.com/jsch/examples/Logger.java.html
-    public static class MyLogger implements com.jcraft.jsch.Logger {
-        static java.util.Hashtable<Integer,String> name=new java.util.Hashtable<>();
-        static{
-            name.put(DEBUG, "DEBUG: ");
-            name.put(INFO, "INFO: ");
-            name.put(WARN, "WARN: ");
-            name.put(ERROR, "ERROR: ");
-            name.put(FATAL, "FATAL: ");
-        }
-        public boolean isEnabled(int level){
-            return true;
-        }
-        public void log(int level, String message){
-            System.err.print(name.get(level));
-            System.err.println(message);
-        }
-    }
-
-
     @Test
     public void testSshPrivateKey() throws Exception {
 
@@ -199,55 +180,23 @@ public class SshPrivateKeyPasswordCloneFXTest extends ApplicationTest {
         console.info("firstCommit name = " + firstCommit.getName());
 
         // Uncomment this to get detail SSH logging info, for debugging
-        // JSch.setLogger(new MyLogger());
+//         JSch.setLogger(new DetailedSshLogger());
 
         // Set up test SSH server.
         try (SshServer sshd = SshServer.setUpDefaultServer()) {
 
-            // Provide SSH server with public and private key info that client will be connecting with
+            String remoteURL = TestUtilities.setUpTestSshServer(sshd,
+                                                                directoryPath,
+                                                                testingRemoteAndLocalRepos.getRemoteFull(),
+                                                                testingRemoteAndLocalRepos.getRemoteBrief());
+
+            console.info("Connecting to " + remoteURL);
+
             InputStream passwordFileStream = getClass().getResourceAsStream("/rsa_key1_passphrase.txt");
             Scanner scanner = new Scanner(passwordFileStream);
             String passphrase = scanner.next();
-            console.info("phrase is " + passphrase);
-
             String privateKeyFileLocation = "/rsa_key1";
-            InputStream privateKeyStream = getClass().getResourceAsStream(privateKeyFileLocation);
-            FilePasswordProvider filePasswordProvider = FilePasswordProvider.of(passphrase);
-            KeyPair kp = SecurityUtils.loadKeyPairIdentity("testkey", privateKeyStream, filePasswordProvider);
-            ArrayList<KeyPair> pairs = new ArrayList<>();
-            pairs.add(kp);
-            KeyPairProvider hostKeyProvider = new MappedKeyPairProvider(pairs);
-            sshd.setKeyPairProvider(hostKeyProvider);
-
-            // Need to use a non-standard port, as there may be an ssh server already running on this machine
-            sshd.setPort(2222);
-
-            // Set up a fall-back password authenticator to help in diagnosing failed test
-            sshd.setPasswordAuthenticator(new PasswordAuthenticator() {
-                public boolean authenticate(String username, String password, ServerSession session) {
-                    fail("Tried to use password instead of public key authentication");
-                    return false;
-                }
-            });
-
-            // This replaces the role of authorized_keys.
-            Collection<PublicKey> allowedKeys = new ArrayList<>();
-            allowedKeys.add(kp.getPublic());
-            sshd.setPublickeyAuthenticator(new KeySetPublickeyAuthenticator(allowedKeys));
-
-            // Amazingly useful Git command setup provided by Mina.
-            sshd.setCommandFactory(new GitPackCommandFactory(directoryPath.toString()));
-
-            // Start the SSH test server.
-            sshd.start();
-
-            // Create temporary known_hosts file.
             Path knownHostsFileLocation = directoryPath.resolve("testing_known_hosts");
-            Files.createFile(knownHostsFileLocation);
-
-            // Clone the bare repo, using the SSH connection, to the local.
-            String remoteURL = "ssh://localhost:2222/" + testingRemoteAndLocalRepos.getRemoteBrief();
-            console.info("Connecting to " + remoteURL);
 
             clickOn("#loadNewRepoButton")
                     .clickOn("#cloneOption")
